+++ src/Instrumentum_sanae_doctrinae/web_scraping/my_constants.py
@@ -69,6 +69,7 @@ BY_SCRIPTURE_ROOT_FOLDER = "by_scripture"
 SCRIPTURE_NAME = "scripture"
 
 BY_SERIES_ROOT_FOLDER = "by_serie"
+SERIE_NAME = "serie"
