+++ tests/monergism/test_monergism_scrap_list.py
@@ -9,10 +9,10 @@ import unittest
 import logging
 logging.basicConfig(level=logging.DEBUG)
 
-from Instrumentum_sanae_doctrinae.scraping import monergism_scrap_metadata
-from Instrumentum_sanae_doctrinae.scraping import monergism_scrap_general_information
-from Instrumentum_sanae_doctrinae.scraping import monergism_scrap_get_list  
-from Instrumentum_sanae_doctrinae.scraping import monergism_scrap_works
+from Instrumentum_sanae_doctrinae.web_scraping import monergism_scrap_metadata
+from Instrumentum_sanae_doctrinae.web_scraping import monergism_scrap_general_information
+from Instrumentum_sanae_doctrinae.web_scraping import monergism_scrap_get_list  
+from Instrumentum_sanae_doctrinae.web_scraping import monergism_scrap_works
 
 
 def test_topic():
